Fix for next step in EPSD100235988, "1.0.0.887 - event log tracking who did what when to what and what was affected - CB event log as an example".
This change adds a log category called actions. Any sam type log that contains
the string "ACTION: " at the beginning of the log message is changed from the
sam category to the action category. I also parse the log message and remove
the "ACTION: " string from the beginning of the message. Additionally, I have
added an "action" category in the GUI advanced selections, but this will not
show up until Nathan or Alex do another GUI build.

To verify that this is working, first ensure you have had some recent actions.
They should be shown in the dashboard area that was labeled events, but should
be soon changed to actions. Then go to the logs tab and search of logs with
the category equal to action. You can used the advanced category tab for this,
or my not so user friendly alternative described below.

If the action category has not yet been added to the advanced log searching,
you can still do it. Just select a log and click on the magnifying glass next
to the category. This will add a filter based upon that category. You can then
go to the filtering section and edit the filter. Click edit on the filter you
created for the category and edit the query to say action instead of whatevr
category you selected in the log.