Ensure scuUnalloacted is accurate
Closes #EPSD100245430, closes #163

Solution is to subtract system.allocated from system.max. We are already using hardcoded zero for all min values concerning utilization. The reason unallocated can be negative with the defect dataset is because scuUnallocated is calculated by summing all the cgroup maxes, then getting the difference from the utilization.scu.system.max. With this data, it results in a negative scuUnallocated value.