Remove controller property changes on scheduler action
Closes #EPSD100244909, closes #122

When taking the schedule action on a node, the schedulerMark property is entirely absent from the sideloaded node. This is the property we rely on for showing a node as scheduled for vm placement (via green highlight). I have assigned the defect to the Poland team for ensuring that the current acted upon node, and the previously selected node are sideloaded with updated properties.